requirements.txt
google-generativeai
Pillow
gTTS
langdetect
playsound
opencv-python
SpeechRecognition
Flask
requests
python-dotenv
